requirements.txt
flask
pillow
pandas
openpyxl
gunicorn